Seminar_1 task_1 commit #004_1

+++ Seminar_1/src/main/java/github/sh1rsh1n/MainApp.java
@@ -10,15 +10,15 @@ public class MainApp {
      /**
       * @param input источник данных(реализующий интерфейс Input)
       * @param handler обработчик данных(реализующий интерфейс NumberHandler), переданных через параметр input
-      * @return int
+      * @return long
       */
-    static int getResult(Input input, NumberHandler<? extends Number> handler) {
-        return handler.operation(input.getInput()).intValue();
+    static long getResult(Input input, NumberHandler<? extends Number> handler) {
+        return handler.operation(input.getInput()).longValue();
     }
 
     public static void main(String[] args) {
 
-        int result = getResult(new ConsoleInput(), new GettingAverageNumber());
+        long result = getResult(new ConsoleInput(), new GettingAverageNumber());
         System.out.printf("Среднее значение: %d\n", result);
     }
 }
